Arrange all the files
