+++ backup/convergence_plots/100_50/merge_20k_0.png
