+++ backup/xor_result.txt
