+++ hmc/make_moon.py
@@ -1,66 +1,35 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# In[1]:
-
-
 import tensorflow as tf
 import tensorflow_probability as tfp
 import numpy as np
 import random
 import time
+import matplotlib.pyplot as plt
+
 from tensorflow import keras
 from tensorflow.keras import layers
 from tensorflow.keras import initializers
 from tensorflow.keras import Model
-from tensorflow.keras.layers import Flatten, Dense
+from tensorflow.keras import models
+from tensorflow.keras.layers import Dense, Conv2D, Dropout, Flatten, MaxPooling2D
+from tensorflow.keras.models import Sequential
 from sklearn.metrics import accuracy_score
+from sklearn.datasets import make_moons
+from sklearn.model_selection import train_test_split
 
-
-# In[2]:
-
-
-# Load MNIST
-(x_train, y_train), (x_test, y_test) = tf.keras.datasets.mnist.load_data()
-
-# Select binary data
-label_sub = [0,1]
-x_train_sub = [x for x, y in zip(x_train, y_train) if y in label_sub]
-y_train_sub = [y for y in y_train if y in label_sub]
-x_test_sub = [x for x, y in zip(x_test, y_test) if y in label_sub]
-y_test_sub = [y for y in y_test if y in label_sub]
-
-
-print('There are', len(x_train_sub), 'training images.')
-print('There are', len(x_test_sub), 'test images.')
-
-
-# In[3]:
-
-
-train_ds = tf.data.Dataset.from_tensor_slices((x_train_sub, y_train_sub)).shuffle(10000).batch(32)
-test_ds = tf.data.Dataset.from_tensor_slices((x_test_sub, y_test_sub)).batch(32)
-
-
-# In[4]:
-
-
+# HMC 
 def convert2_zero_one(x):
-
-    t = [tf.math.sigmoid(i) for i in x]
     
+    t = [tf.math.sigmoid(i) for i in x]    
     return t
 
 def rerange(x, r = 6.0):
     
     out_of_range = tf.cast(tf.math.greater(tf.math.abs(x), r), tf.float32)
     sign = tf.math.sign(x)
-    
     return x * (1 - out_of_range) + sign * r * out_of_range
-
-
-# In[5]:
-
 
 # MLP model
 class StochasticMLP(Model):
@@ -72,8 +41,9 @@ class StochasticMLP(Model):
         self.output_layer = Dense(n_outputs)
     
     def call(self, x):
-
-        x = Flatten()(x)
+        
+        #x = Flatten()(x)
+        
         network = []
         
         for i, layer in enumerate(self.fc_layers):
@@ -88,8 +58,6 @@ class StochasticMLP(Model):
     
     def target_log_prob(self, x, h, y):
         
-        x = Flatten()(x)
-        y = [[i] for i in y]
         h_current = convert2_zero_one([tf.cast(h_i, dtype=tf.float32) for h_i in h])
         h_previous = [x] + h_current[:-1]
     
@@ -110,11 +78,8 @@ class StochasticMLP(Model):
         return -1 * nlog_prob
 
     def target_log_prob2(self, x, h, y):
-        
-        x = Flatten()(x)
-        y = [[i] for i in y]
+
         h_current = convert2_zero_one(tf.split(h, self.hidden_layer_sizes, axis = 1))
-        #h_current = [h_current[0]]
         h_previous = [x] + h_current[:-1]
         
         nlog_prob = 0.
@@ -144,34 +109,31 @@ class StochasticMLP(Model):
         return adaptive_hmc
     
     # new proposing-state method with HamiltonianMonteCarlo
-    def propose_new_state_hamiltonian(self, x, h, y, hmc_ker):
+    def propose_new_state_hamiltonian(self, x, h, y, hmc_kernel):
     
         h_current = h
         h_current = [tf.cast(h_i, dtype=tf.float32) for h_i in h_current]
-        h_current = tf.concat([h[0], h[1]], axis=1)
-
-        # initialize the HMC transition kernel
-        
-        adaptive_hmc = tfp.mcmc.SimpleStepSizeAdaptation(tfp.mcmc.HamiltonianMonteCarlo(
-            target_log_prob_fn = lambda v: self.target_log_prob2(x, v, y),
-            num_leapfrog_steps = 2,
-            step_size = pow(1000, -1/4)),
-            num_adaptation_steps=int(100*0.8))
+        h_current = tf.concat(h_current, axis = 1)
 
         # run the chain (with burn-in)
+        num_burnin_steps = 0
         num_results = 1
-        num_burnin_steps = 100
 
         samples = tfp.mcmc.sample_chain(
             num_results = num_results,
             num_burnin_steps = num_burnin_steps,
             current_state = h_current, # may need to be reshaped
-            kernel = adaptive_hmc,
-            trace_fn = None)
+            kernel = hmc_kernel,
+            trace_fn = None,
+            return_final_kernel_results = True)
+        
+        new_step_size = samples[2].new_step_size.numpy()
+        ker_new = self.generate_hmc_kernel(x, y, new_step_size)
+        
+        h_state = rerange(samples[0][0])
+        h_new = tf.split(h_state, self.hidden_layer_sizes, axis = 1)   
 
-        h_new = tf.split(samples[0], self.hidden_layer_sizes, axis = 1)
-
-        return(h_new)
+        return(h_new, ker_new)
     
     def update_weights(self, x, h, y, lr = 0.1):
         
@@ -184,7 +146,6 @@ class StochasticMLP(Model):
 
     def get_predictions(self, x):
 
-        x = Flatten()(x)
         logits = 0.0
         for layer in self.fc_layers:
             logits = layer(x)
@@ -197,111 +158,59 @@ class StochasticMLP(Model):
 
         return labels
 
+# Make data
+np.random.seed(1234)
+X, Y = make_moons(500, noise = 0.1)
 
-# In[14]:
+# Split into test and training data
+x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.25, random_state=73)
+y_train = y_train.reshape((-1, 1))
+y_test = y_test.reshape((-1, 1))
 
+train_ds = tf.data.Dataset.from_tensor_slices((x_train, y_train)).batch(32)
+test_ds = tf.data.Dataset.from_tensor_slices((x_test, y_test)).batch(32)
 
-model = StochasticMLP(hidden_layer_sizes = [100, 50], n_outputs = 1)
+model = StochasticMLP(hidden_layer_sizes = [100], n_outputs=1)
+network = [model.call(x) for x, y in train_ds]
+kernels = [model.generate_hmc_kernel(x, y) for x, y in train_ds]
 
-
-# In[15]:
-
-
-network = [model.call(images) for images, labels in train_ds]
-
-
-# In[16]:
-
-
-kernels = [model.generate_hmc_kernel(images, labels) for images, labels in train_ds]
-
-
-# In[18]:
-
-
-burnin = 20
+# Burn-in
+burnin = 10
 step_sizes = []
 for i in range(burnin):
     
-    #print(i)
+    #if(i % 100 == 0):
+    print("Step %d" % i)
     network_new = []
     kernels_new = []
-    start_time = time.time()
     
-    for (images, labels), net, hmc_kernel in zip(train_ds, network, kernels):
-        net_current = net
-        net_current = [tf.cast(net_i, dtype=tf.float32) for net_i in net_current]
-        #net_current = tf.concat([net_current[0], net_current[1]], axis=1)
-        net_current = tf.concat(net_current, axis = 1)
-        
-        num_results = 1
-        num_burnin_steps = 0
-
-        samples = tfp.mcmc.sample_chain(
-            num_results = num_results,
-            num_burnin_steps = num_burnin_steps,
-            current_state = net_current, # may need to be reshaped
-            kernel = hmc_kernel,
-            #trace_fn = lambda _, pkr: pkr.inner_results.accepted_results.new_step_size,
-            trace_fn = None,
-            return_final_kernel_results = True)
-        
-        #print(samples[2].new_step_size.numpy())
-        new_step_size = samples[2].new_step_size.numpy()
-        step_sizes.append(new_step_size)
-
-        new_state = rerange(samples[0][0])
-        net_new = tf.split(new_state, [100, 50], axis = 1)   
-        network_new.append(net_new)
-        
-        # build new kernel
-        ker_new = model.generate_hmc_kernel(images, labels, new_step_size)
-        kernels_new.append(ker_new)
-            
+    res = [model.propose_new_state_hamiltonian(x, net, y, ker) 
+               for (x, y), net, ker in zip(train_ds, network, kernels)]
+    
+    network_new, kernels_new = zip(*res)
+         
     network = network_new
     kernels = kernels_new
-    
-    print("Step %d - time %.4f" % (i, time.time()-start_time))
 
-
-# In[ ]:
-
-
-print(network[0])
-
-
-# In[ ]:
-
-
-import matplotlib.pyplot as plt
-plt.plot(np.arange(len(step_sizes)), step_sizes)
-plt.show()
-
-
-# In[ ]:
-
-
-epochs = 500
-
+# Training
+epochs = 1000
 start_time = time.time()
 for epoch in range(epochs):
     
     loss = 0.0
     acc = 0.0
-    for bs, (images, labels) in enumerate(train_ds):
+    for bs, (x, y) in enumerate(train_ds):
         
         # only one mini-batch
-        model.update_weights(images, network[bs], labels, 0.1)
-        #network_new = [model.propose_new_state_hamiltonian(images, net, labels) for (images, labels), net in 
-        #               zip(train_ds, network)]
+        model.update_weights(x, network[bs], y, 0.1)
         
         network_new = []
-        #kernels_new = []
-        for net, hmc_kernel in zip(network, kernels):
+
+        for (xx, yy), net, hmc_kernel in zip(train_ds, network, kernels):
             net_current = net
             net_current = [tf.cast(net_i, dtype=tf.float32) for net_i in net_current]
             net_current = tf.concat(net_current, axis = 1)
-            
+        
             num_results = 1
             num_burnin_steps = 0
 
@@ -312,30 +221,17 @@ for epoch in range(epochs):
                 kernel = hmc_kernel,
                 trace_fn = None,
                 return_final_kernel_results = True)
-
-            #new_step_size = samples[2].new_step_size.numpy()
             
             new_state = rerange(samples[0][0])
-            net_new = tf.split(new_state, [100, 50], axis = 1)   
+            net_new = tf.split(new_state, [100], axis = 1)   
             network_new.append(net_new)
-            
-            #ker_new = model.generate_hmc_kernel(images2, labels2, new_step_size)
-            #kernels_new.append(ker_new)
-            
+   
         network = network_new
-        #kernels = kernels_new
-
-        loss += -1 * tf.reduce_mean(model.target_log_prob(images, network[bs], labels))
-       
+        
+        loss += -1 * tf.reduce_mean(model.target_log_prob(x, network[bs], y))
+    
     preds = [model.get_predictions(images) for images, labels in train_ds]
-    #print(preds)
-    train_acc = accuracy_score(np.array(preds[0]), y_train)
+    train_acc = accuracy_score(np.concatenate(preds), y_train)
     print("Epoch %d/%d: - %.4fs/step - loss: %.4f - accuracy: %.4f" 
           % (epoch + 1, epochs, (time.time() - start_time) / (epoch + 1), loss, train_acc))
-
-
-# In[ ]:
-
-
-
-
+'''
