+++ hmc/make_moon_mpi.py
@@ -1,0 +1,238 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+import tensorflow as tf
+import tensorflow_probability as tfp
+import numpy as np
+import random
+import time
+import matplotlib.pyplot as plt
+
+from tensorflow import keras
+from tensorflow.keras import layers
+from tensorflow.keras import initializers
+from tensorflow.keras import Model
+from tensorflow.keras import models
+from tensorflow.keras.layers import Dense, Conv2D, Dropout, Flatten, MaxPooling2D
+from tensorflow.keras.models import Sequential
+from sklearn.metrics import accuracy_score
+from sklearn.datasets import make_moons
+from sklearn.model_selection import train_test_split
+
+from mpi4py import MPI
+
+comm = MPI.COMM_WORLD
+size = comm.Get_size()
+rank = comm.Get_rank()
+
+ndata = 1000
+numPerRank = int(np.floor((ndata * 0.8) / size))
+
+x_train = None
+y_train = None
+x_test = None
+y_test = None
+
+# Make data
+if rank == 0:
+    np.random.seed(1234)
+    X, Y = make_moons(ndata, noise = 0.1)
+    x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.2, random_state=73)
+    y_train = y_train.reshape((-1, 1))
+    y_test = y_test.reshape((-1, 1))
+
+x_trn = np.empty((numPerRank, 2), dtype = "float64")
+y_trn = np.empty((numPerRank, 1), dtype = "int")
+comm.Scatter(x_train, x_trn, root = 0)
+comm.Scatter(y_train, y_trn, root = 0)
+
+train_ds = tf.data.Dataset.from_tensor_slices((x_trn, y_trn)).batch(32)
+
+# HMC 
+def convert2_zero_one(x):
+    
+    t = [tf.math.sigmoid(i) for i in x]    
+    return t
+
+def rerange(x, r = 6.0):
+    
+    out_of_range = tf.cast(tf.math.greater(tf.math.abs(x), r), tf.float32)
+    sign = tf.math.sign(x)
+    return x * (1 - out_of_range) + sign * r * out_of_range
+
+# MLP model
+class StochasticMLP(Model):
+    
+    def __init__(self, hidden_layer_sizes=[100], n_outputs=10):
+        super(StochasticMLP, self).__init__()
+        self.hidden_layer_sizes = hidden_layer_sizes
+        self.fc_layers = [Dense(layer_size) for layer_size in hidden_layer_sizes]
+        self.output_layer = Dense(n_outputs)
+    
+    def call(self, x):
+        
+        #x = Flatten()(x)
+        
+        network = []
+        
+        for i, layer in enumerate(self.fc_layers):
+            
+            logits = layer(x)
+            x = tfp.distributions.Bernoulli(logits=logits).sample()
+            network.append(x)
+
+        final_logits = self.output_layer(x) # initial the weight of output layer
+            
+        return network
+    
+    def target_log_prob(self, x, h, y):
+        
+        h_current = convert2_zero_one([tf.cast(h_i, dtype=tf.float32) for h_i in h])
+        h_previous = [x] + h_current[:-1]
+    
+        nlog_prob = 0. # negative log probability
+        
+        for i, (cv, pv, layer) in enumerate(
+            zip(h_current, h_previous, self.fc_layers)):
+            
+            ce = tf.nn.sigmoid_cross_entropy_with_logits(
+                labels=cv, logits=layer(pv))
+            
+            nlog_prob += tf.reduce_sum(ce, axis = -1)
+        
+        fce = tf.nn.sigmoid_cross_entropy_with_logits(
+            labels=tf.cast(y, tf.float32), logits=self.output_layer(h_current[-1]))
+        nlog_prob += tf.reduce_sum(fce, axis = -1)
+            
+        return -1 * nlog_prob
+
+    def target_log_prob2(self, x, h, y):
+
+        h_current = convert2_zero_one(tf.split(h, self.hidden_layer_sizes, axis = 1))
+        h_previous = [x] + h_current[:-1]
+        
+        nlog_prob = 0.
+        
+        for i, (cv, pv, layer) in enumerate(
+            zip(h_current, h_previous, self.fc_layers)):
+            
+            ce = tf.nn.sigmoid_cross_entropy_with_logits(
+                labels=cv, logits=layer(pv))
+            
+            nlog_prob += tf.reduce_sum(ce, axis = -1)
+        
+        fce = tf.nn.sigmoid_cross_entropy_with_logits(
+            labels=tf.cast(y, tf.float32), logits=self.output_layer(h_current[-1]))
+        nlog_prob += tf.reduce_sum(fce, axis = -1)
+            
+        return -1 * nlog_prob
+    
+    def generate_hmc_kernel(self, x, y, step_size = pow(1000, -1/4)):
+        
+        adaptive_hmc = tfp.mcmc.SimpleStepSizeAdaptation(tfp.mcmc.HamiltonianMonteCarlo(
+            target_log_prob_fn = lambda v: self.target_log_prob2(x, v, y),
+            num_leapfrog_steps = 2,
+            step_size = step_size),
+            num_adaptation_steps=int(100 * 0.8))
+        
+        return adaptive_hmc
+    
+    # new proposing-state method with HamiltonianMonteCarlo
+    def propose_new_state_hamiltonian(self, x, h, y, hmc_kernel):
+    
+        h_current = h
+        h_current = [tf.cast(h_i, dtype=tf.float32) for h_i in h_current]
+        h_current = tf.concat(h_current, axis = 1)
+
+        # run the chain (with burn-in)
+        num_burnin_steps = 0
+        num_results = 1
+
+        samples = tfp.mcmc.sample_chain(
+            num_results = num_results,
+            num_burnin_steps = num_burnin_steps,
+            current_state = h_current, # may need to be reshaped
+            kernel = hmc_kernel,
+            trace_fn = None,
+            return_final_kernel_results = True)
+        
+        new_step_size = samples[2].new_step_size.numpy()
+        ker_new = self.generate_hmc_kernel(x, y, new_step_size)
+        
+        h_state = rerange(samples[0][0])
+        h_new = tf.split(h_state, self.hidden_layer_sizes, axis = 1)   
+
+        return(h_new, ker_new)
+    
+    def update_weights(self, x, h, y, lr = 0.1):
+        
+        optimizer = tf.keras.optimizers.SGD(learning_rate = lr)
+        with tf.GradientTape() as tape:
+            loss = -1 * tf.reduce_mean(self.target_log_prob(x, h, y))
+        
+        grads = tape.gradient(loss, self.trainable_weights)
+        optimizer.apply_gradients(zip(grads, self.trainable_weights))
+
+    def get_predictions(self, x):
+
+        logits = 0.0
+        for layer in self.fc_layers:
+            logits = layer(x)
+            x = tf.math.sigmoid(logits)
+        
+        logits = self.output_layer(x)
+        probs = tf.math.sigmoid(logits)
+        #print(probs)
+        labels = tf.cast(tf.math.greater(probs, 0.5), tf.int32)
+
+        return labels
+
+model = StochasticMLP(hidden_layer_sizes = [50], n_outputs=1)
+network = [model.call(x) for x, y in train_ds]
+kernels = [model.generate_hmc_kernel(x, y) for x, y in train_ds]
+
+print("Rank: ", rank, "Network: ", len(network))
+print("Rank: ", rank,)
+
+# Burn-in
+burnin = 500
+for i in range(burnin):
+    
+    if(i % 100 == 0):
+        print("Step %d" % i)
+    network_new = []
+    kernels_new = []
+    
+    res = [model.propose_new_state_hamiltonian(x, net, y, ker) 
+                for (x, y), net, ker in zip(train_ds, network, kernels)]
+    
+    network_new, kernels_new = zip(*res)
+    network = network_new
+    kernels = kernels_new
+
+#print("Rank: ", rank, "Network: ", network[0])
+
+# Training
+epochs = 1000
+start_time = time.time()
+for epoch in range(epochs):
+    
+    loss = 0.0
+    acc = 0.0
+    for bs, (x, y) in enumerate(train_ds):
+        
+        # only one mini-batch
+        model.update_weights(x, network[bs], y, 0.1)
+        
+        network_new = []
+        res = [model.propose_new_state_hamiltonian(x, net, y, ker)
+                    for (x, y), net, ker in zip(train_ds, network, kernels)]
+        network = network_new
+        
+        loss += -1 * tf.reduce_mean(model.target_log_prob(x, network[bs], y))
+    
+    preds = [model.get_predictions(images) for images, labels in train_ds]
+    train_acc = accuracy_score(np.concatenate(preds), y_train)
+    print("Epoch %d/%d: - %.4fs/step - loss: %.4f - accuracy: %.4f" 
+          % (epoch + 1, epochs, (time.time() - start_time) / (epoch + 1), loss, train_acc))
+'''
